test(generate_enum_test): organize generator tests
