+++ test/bin/src/generators/generate_enum_test.dart
@@ -6,7 +6,7 @@ import 'package:path/path.dart' as path;
 import 'package:supabase_codegen/supabase_codegen.dart';
 import 'package:test/test.dart';
 
-import '../../../bin/src/src.dart';
+import '../../../../bin/src/src.dart';
 
 void main() {
   group('generateEnumsFile', () {
